Added swedish date format + weeks on datepickers
